requirements.txt
Flask
itsdangerous
flask-login
Flask-SQLAlchemy
psycopg2
requests
gunicorn
gevent
marshmallow
sqlparse
pdfkit
pikepdf
PyPDF2
openpyxl
XlsxWriter
Pillow
terbilang
blinker
pyjwt